requirements.txt
# requirements.txt
# Core Qiskit Components
qiskit>=1.0
qiskit_aer>=0.13 # Use the new standalone Aer package

# Scientific Computing
numpy>=1.20
scipy>=1.8

# Configuration/Data Files
PyYAML>=5.4

# Optional: For Examples/Visualization
matplotlib>=3.5

# Optional: For Development/Testing
# pytest>=7.0
# flake8>=5.0